Fix JsonDecodingException by making bpm a Double

+++ app/src/main/java/fr/m335/subtide/data/SubwaveApi.kt
@@ -22,7 +22,7 @@ data class NowPlayingTrack(
     val album: String? = null,
     @SerialName("subsonic_id") val subsonicId: String? = null,
     val genre: String? = null,
-    val bpm: Int? = null,
+    val bpm: Double? = null,
     val musicalKey: String? = null,
     val moods: List<String>? = null,
     val year: Int? = null,
